+++ build.gradle
@@ -61,6 +61,10 @@ dependencies {
     annotationProcessor "com.querydsl:querydsl-apt:${dependencyManagement.importedProperties['querydsl.version']}:jakarta"
     annotationProcessor "jakarta.annotation:jakarta.annotation-api"
     annotationProcessor "jakarta.persistence:jakarta.persistence-api"
+
+    // 인메모리 몽고디비 flapdoodle
+    testImplementation 'de.flapdoodle.embed:de.flapdoodle.embed.mongo:4.6.0'
+
 }
 
 tasks.named('test') {
