--- src/main/java/com/example/api/account/AccountRepository.java
@@ -1,9 +1,0 @@
-package com.example.api.account;
-
-import com.example.api.domain.Account;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-interface AccountRepository extends JpaRepository<Account, Long> {
-}
