--- src/main/java/com/example/api/account/AccountService.java
@@ -1,17 +1,0 @@
-package com.example.api.account;
-
-import com.example.api.domain.Account;
-import com.example.api.exception.BusinessException;
-import com.example.api.exception.ErrorCode;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class AccountService {
-    private final AccountRepository accountRepository;
-    public Account loadAccount(final Long requestMemberId) {
-        return accountRepository.findById(requestMemberId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.ACCOUNT_NOT_FOUND_EXCEPTION));
-    }
-}
