+++ src/main/java/com/example/api/board/service/BoardService.java
@@ -3,7 +3,11 @@ package com.example.api.board.service;
 
 import com.example.api.board.controller.domain.request.EmployeeIdRequest;
 import com.example.api.board.controller.domain.response.MyInfoDTO;
-import com.example.api.domain.repository.*;
+import com.example.api.domain.repository.ExternalCareerRepository;
+import com.example.api.domain.repository.FlavoredRepository;
+import com.example.api.domain.repository.MyInfoRepository;
+import com.example.api.domain.repository.OfferEmploymentRepository;
+import com.example.api.possbileboard.PossibleBoardRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
