+++ src/main/resources/application.properties
@@ -23,11 +23,13 @@ logging.level.org.hibernate.SQL=DEBUG
 logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
 
 # Mongo Database Configuration
+spring.data.mongodb.uri=mongodb://localhost:27017/testdb
+
 spring.data.mongodb.host=localhost
-spring.data.mongodb.port=27017
-spring.data.mongodb.authentication-database=admin
-spring.data.mongodb.database=danpatime
-spring.data.mongodb.auto-index-creation=true
+#spring.data.mongodb.port=27017
+#spring.data.mongodb.authentication-database=admin
+#spring.data.mongodb.database=testdb
+#spring.data.mongodb.auto-index-creation=true
 
 # SMTP
 spring.mail.host=smtp.gmail.com
